Move `IpEndpoint` to `socket/ip/addr.rs`
